+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the "tswagger" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [Unreleased]
+
+### Added
+
+- Support to map the `basePath` by adding `basePathMapping` config.
+
 ## [1.1.1] - 2023-12-17
 
 ### Fixed